feat(eslint): add rule to restrict supabase imports to infrastructure layer
Add ESLint rule that restricts imports from '~/shared/utils/supabase' to infrastructure layer only, enforcing clean architecture principles.

Changes:
- Add no-restricted-imports rule for supabase utils in eslint.config.mjs
- Allow supabase imports only in infrastructure/** files and specific exceptions
- Move search cache functions from application to infrastructure layer
- Update search application layer to re-export infrastructure functions
- Ensure all quality checks pass with new architectural constraints

This prevents application and domain layers from directly accessing Supabase,
requiring proper use of repository abstractions.
